frontend progress: posts by brand, improve user ui

+++ core/views/posts.py
@@ -1,5 +1,5 @@
 
-from core.serializers import CustomUserSerializer, PostSerializer, FollowSerializer, CommentSerializer#, GroupSerializer
+from core.serializers import CustomUserSerializer, PostSerializer, PostSerializerFull, FollowSerializer, CommentSerializer#, GroupSerializer
 from django.http.response import JsonResponse
 from rest_framework import status
 
@@ -7,6 +7,8 @@ from core.models import CustomUser, Post, Follow, Item, Media, MediaItem, PostLi
 from rest_framework.decorators import api_view
 
 from .users import get_user_by_id
+
+import json
 
    
 def get_post_by_id(id):
@@ -39,7 +41,7 @@ def post_detail(request, id):
     return JsonResponse(postSerializer.data, status=status.HTTP_200_OK)
 
 
-# TODO: no limit on this one - for now.
+# TODO: limit + paginate this. Sort those users somehow.
 @api_view(['GET'])
 def post_likes(request, id):
     # gets the users who have liked this post
@@ -55,8 +57,18 @@ def post_likes(request, id):
     userSerializer = UserSerializer(users, many=True)
     return JsonResponse(userSerializer.data, safe=False, status=status.HTTP_200_OK)
 
+@api_view(['GET'])
+def post_num_likes(request, id):
+    # gets the post's number of likes
 
-@api_view(['POST'])
+    post = get_post_by_id(id)
+    if post is None:
+        return JsonResponse({'message': 'The post does not exist'}, status=status.HTTP_404_NOT_FOUND) 
+
+    return JsonResponse({'num_likes': post.num_likes}, status=status.HTTP_200_OK)
+
+
+@api_view(['GET','POST'])
 def post_like(request, id, userid):
     # likes/unlikes a post by postid for userid
 
@@ -68,36 +80,46 @@ def post_like(request, id, userid):
     if user is None:
         return JsonResponse({'message': 'The user does not exist'}, status=status.HTTP_404_NOT_FOUND) 
     
-    like = PostLike.objects.filter(post=post, user=user)
-    if like: #like already exists -> decrement it
-        like.delete()
-        post.decrement_likes()
-    else:
-        PostLike.create(post=post, user=user)
-        post.increment_likes()
+    like = PostLike.objects.filter(post=post, user=user).first()
 
-    return JsonResponse({'message': 'success'}, status=status.HTTP_201_CREATED)
+    if request.method == 'GET':
+        return JsonResponse({'status': like is not None}, status=status.HTTP_200_OK)
+    
+    elif request.method == 'POST':
+        if like: #like already exists -> decrement it
+            like.delete()
+            post.decrement_likes()
+        else:
+            PostLike.objects.create(post=post, user=user)
+            post.increment_likes()
+
+        return JsonResponse({'message': 'success'}, status=status.HTTP_201_CREATED)
 
 
 MAX_COMMENT_LENGTH = 500
 
 @api_view(['POST'])
-def post_comment(request, id, userid):
+def post_comment(request, id):
     # comments a post by postid for userid
 
-    content = request.POST.get("data")
-    if not data or data == "" or len(data) > MAX_COMMENT_LENGTH:
+    data = json.loads(request.body)
+
+    content = data.get("content")
+    if not content or content == "" or len(content) > MAX_COMMENT_LENGTH:
         return JsonResponse({'message': 'Invalid comment'}, status=status.HTTP_400_BAD_REQUEST)
 
     post = get_post_by_id(id)
     if post is None:
         return JsonResponse({'message': 'The post does not exist'}, status=status.HTTP_404_NOT_FOUND) 
 
+    userid = data.get("user")
+    if not userid:
+        return JsonResponse({'message': 'Invalid comment'}, status=status.HTTP_400_BAD_REQUEST)
     user = get_user_by_id(userid)
     if user is None:
         return JsonResponse({'message': 'The user does not exist'}, status=status.HTTP_404_NOT_FOUND) 
 
-    Comment.create(
+    Comment.objects.create(
         post=post,
         parent_comment=None,
         user=user,
@@ -127,18 +149,15 @@ def post_comments(request, id):
 
 @api_view(['GET'])
 def posts_containing_item_brand(request, brand):
-    # may have to do: [i.mediaItems for i in Item.objects.filter...]
-    # and then mediaItem.post for mediaItem in that ^
-    # TODO: yeah this is wrong v
-    # item filter -> mediaItems -> media -> post
-    # or item filter -> media (via media_set I think)
-    posts = Item.objects.filter(brand=brand).media_set.all().post.all()
+    posts = Post.objects.filter(mediaItems__item__brand=brand).distinct()
 
+    """
     order = request.params.get("order")
     if not order or order == "most recent":
         posts = posts.order_by('-timestamp')
     elif order == "likes":
         posts = posts.order_by('-num_likes')
+    """
 
-    postSerializer = PostSerializer(posts, many=True)
+    postSerializer = PostSerializerFull(posts, many=True)
     return JsonResponse(postSerializer.data, safe=False, status=status.HTTP_200_OK)
